requirements.txt

google-search-results==2.4.2
langchain==0.0.316
langcodes==3.3.0
langsmith==0.0.43
numexpr==2.8.7
openai==0.27.10
pydantic==1.10.13
pydantic_core==2.10.1
python-dotenv==1.0.0
streamlit==1.27.2
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-chat==0.1.1
streamlit-embedcode==0.1.2
streamlit-extras==0.3.4
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
wikipedia==1.4.0